% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/closeAllTerminals.R
\name{closeAllTerminals}
\alias{closeAllTerminals}
\title{Close all terminals}
\usage{
closeAllTerminals()
}
\description{
Closes all terminals, including busy terminals. Kills any processes running within
the terminals without warning the user.
}
