% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terminaltools-package.R
\docType{package}
\name{terminaltools}
\alias{terminaltools}
\alias{terminaltools-package}
\title{RStudio Terminal Tools Addin}
\description{
An RStudio addin providing additional commands for the RStudio terminal.
}
